Rename files according to their used location lang
