+++ packages/marko-web-leaders/browser/company-website-link.vue
@@ -12,7 +12,7 @@
 </template>
 
 <script>
-import i18n from '../i18n';
+import i18n from '../i18n-vue';
 
 export default {
   props: {
